Subindo novas implementações.

+++ src/modules/turmas/turmas.service.ts
@@ -1,11 +1,10 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { PrismaService } from '../../shared/prisma/prisma.service';
 
 @Injectable()
 export class TurmasService {
   constructor(private readonly prisma: PrismaService) {}
 
-  // Criar uma nova turma
   async createTurma(data: { name: string; professorId: number }) {
     return this.prisma.turma.create({
       data: {
@@ -15,7 +14,6 @@ export class TurmasService {
     });
   }
 
-  // Listar todas as turmas
   async findAllTurmas() {
     return this.prisma.turma.findMany({
       include: {
@@ -25,22 +23,29 @@ export class TurmasService {
     });
   }
 
-  // Atualizar informações de uma turma
   async updateTurma(id: number, data: { name?: string; professorId?: number }) {
+    const turma = await this.prisma.turma.findUnique({ where: { id } });
+    if (!turma) {
+      throw new Error('Turma não encontrada');
+    }
+
     return this.prisma.turma.update({
       where: { id },
       data,
     });
   }
 
-  // Excluir uma turma
   async deleteTurma(id: number) {
+    const turma = await this.prisma.turma.findUnique({ where: { id } });
+    if (!turma) {
+      throw new NotFoundException('Turma não encontrada');
+    }
+
     return this.prisma.turma.delete({
       where: { id },
     });
   }
 
-  // Listar turmas de um professor específico
   async findProfessorTurmas(professorId: number) {
     return this.prisma.turma.findMany({
       where: { professorId },
@@ -48,8 +53,15 @@ export class TurmasService {
     });
   }
 
-  // Adicionar um aluno a uma turma
-  async addAlunoToTurma(turmaId: number, alunoId: number) {
+  async addAlunoToTurma(turmaId: number, alunoId: number, professorId: number) {
+    const turma = await this.prisma.turma.findUnique({
+      where: { id: turmaId },
+    });
+
+    if (!turma || turma.professorId !== professorId) {
+      throw new Error('Acesso negado ou turma não encontrada');
+    }
+
     return this.prisma.turma.update({
       where: { id: turmaId },
       data: {
@@ -60,13 +72,18 @@ export class TurmasService {
     });
   }
 
-  // Listar alunos de uma turma
   async findAlunosByTurma(turmaId: number) {
     return this.prisma.user.findMany({
       where: {
-        turmaId, // Verifica se o aluno pertence à turma
-        role: 'aluno', // Filtra apenas os usuários com o papel de aluno
+        turmaId,
+        role: 'aluno',
       },
+    });
+  }
+
+  async findTurmaById(turmaId: number) {
+    return this.prisma.turma.findUnique({
+      where: { id: turmaId },
     });
   }
 }
